Added timer & graphics
There is now a timer for how long user's have before the game ends. With
this, a button for starting each round was also added.

New graphics and CSS has been added to style the game. Only required bit
left is images for the targets. I'd also like to add time bonuses for
some targets.